// app/types/api-data.ts
export type UserData = {
    avatar: string;
    business_customer: number;
    created_at: string;
    deleted_at: string | null;
    email: string;
    email_verified_at: string | null;
    id: number;
    is_administrator: number;
    last_name: string;
    name: string;
    phone: string;
    phone_verified_at: string | null;
    updated_at: string;
  };
  